+++ lib/generators/rspec/controller/templates/controller_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 <% module_namespacing do -%>
-describe <%= class_name %>Controller do
+describe <%= class_name %>Controller, :type => :controller do
 
 <% for action in actions -%>
   describe "GET '<%= action %>'" do