+++ lib/generators/rspec/controller/templates/view_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe "<%= file_name %>/<%= @action %>.html.<%= options[:template_engine] %>" do
+describe "<%= file_name %>/<%= @action %>.html.<%= options[:template_engine] %>", :type => :view do
   pending "add some examples to (or delete) #{__FILE__}"
 end