+++ spec/generators/rspec/model/model_generator_spec.rb
@@ -27,7 +27,7 @@ describe Rspec::Generators::ModelGenerator, :type => :generator do
 
         it { is_expected.to exist }
         it { is_expected.to contain(/require 'spec_helper'/) }
-        it { is_expected.to contain(/describe Posts/) }
+        it { is_expected.to contain(/describe Posts, :type => :model/) }
       end
 
       describe 'the fixtures' do